#28 Refactor: 데일리 챌린지, chaleengejoin 변경
