+++ public/Chamyo.json
@@ -1,0 +1,20 @@
+{
+  "joinList": [
+    {
+      "images": "url",
+      "nickName": "꿍이집사"
+    },{
+      "images": "url",
+      "nickName": "꿍이집사"
+    },{
+      "images": "url",
+      "nickName": "꿍이집사"
+    },{
+      "images": "url",
+      "nickName": "꿍이집사"
+    },{
+      "images": "url",
+      "nickName": "꿍이집사"
+    }
+  ]
+}
