+++ src/styles/challengejoin.module.scss
@@ -24,7 +24,6 @@
       width: 120px;
       height: 120px;
       border-radius: 12px;
-      background-color: yellowgreen;
       margin-top: 10px;
     }
     .name {
