+++ src/utils/Scroll.ts
@@ -1,0 +1,45 @@
+import { useRef, useEffect } from 'react';
+
+export const useScroll = () => {
+  const scrollRef = useRef<HTMLDivElement>(null);
+  let startTouchX = 0;
+
+  useEffect(() => {
+    const handleWheel = (e:any) => {
+      if (e.deltaY === 0 || !scrollRef.current) return;
+      e.preventDefault();
+      if (scrollRef.current) {
+        scrollRef.current.scrollLeft += e.deltaY;
+      }
+    };
+    const handleTouchStart = (e:any) => {
+      if (!scrollRef.current) return;
+      startTouchX = e.touches[0].clientX; 
+    };
+    const handleTouchMove = (e:any) => {
+      if (!scrollRef.current) return;
+      e.preventDefault();
+      const xDiff = startTouchX - e.touches[0].clientX;
+      if (scrollRef.current){ 
+        scrollRef.current.scrollLeft += xDiff;
+        startTouchX = e.touches[0].clientX;
+      }
+    };
+    const element = scrollRef.current;
+    if (element) {
+      element.addEventListener('wheel', handleWheel, { passive: false });
+      element.addEventListener('touchstart', handleTouchStart);
+     element.addEventListener('touchmove', handleTouchMove);
+
+     return () => {
+       if(element){
+         element.removeEventListener('wheel', handleWheel);
+         element.removeEventListener('touchstart',handleTouchStart);
+         element.removeEventListener('touchmove',handleTouchMove);
+       }
+     };
+   }
+ }, []);
+
+ return scrollRef;
+};
